restructure ssr tests: move template tests into separate file
